+++ packages/compostate/src/core/constants.ts
@@ -1,0 +1,7 @@
+export function IS_EQUAL(a: unknown, b: unknown) {
+  return a === b || (a !== a && b !== b);
+}
+
+export function NO_OP(): void {
+  // no-op
+}
